backend/requirements.txt
fastapi==0.68.3
uvicorn==0.15.0
pydantic==1.9.0
passlib==1.7.5
motor==2.5.1
python-dotenv==0.19.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
